3 api for devices
