+++ src/test/resources/config.properties
@@ -4,28 +4,14 @@ username=admin
 password=password123
 
 ## URL ##
-url = https://restful-booker.herokuapp.com/
+url = https://demo.webmate.io/api/v1/
+projectId = 7d4e5fa6-5075-4220-857f-a4ec4a0cf8fa
 
-## Create Data ##
+getDevices = projects/{projectId}/device/devices
+apiToken = 3ae9783b-8367-49a0-8f17-001385e3b6f5
+deleteDevice = device/devices/{deviceId}
+screenShot=device/{deviceId}/screenshot
 
-firstname =Himanshu
-lastname=K
-totalprice=111
-depositpaid=false,
-bookingdates
-checkin=2018-01-01
-checkout=2019-01-01
-additionalneeds=Breakfast
-
-## Update Data ##
-firstname_u=Himanshu
-lastname_u=Kumar
-totalprice_u=111
-depositpaid_u=false,
-bookingdates
-checkin_u=2018-01-01
-checkout_u=2019-01-01
-additionalneeds_u=Breakfast
-
+webMateUser=aditya.rawat@nagarro.com
 
 urlui=http://demo.guru99.com/test/newtours/
