+++ test-output/Default suite/Default test.html
@@ -75,7 +75,7 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='apitest.DeleteBookingTestApi.InvalidDelete()'><b>InvalidDelete</b><br>Test class: apitest.DeleteBookingTestApi</td>
+<td title='apitest.ReleaseDevice.InvalidDelete()'><b>InvalidDelete</b><br>Test class: apitest.ReleaseDevice</td>
 <td><div><pre>java.lang.AssertionError: 1 expectation failed.
 Expected status code &lt;405&gt; but was &lt;403&gt;.
 
@@ -113,7 +113,7 @@ Expected status code &lt;405&gt; but was &lt;403&gt;.
 	at org.codehaus.groovy.runtime.ScriptBytecodeAdapter.invokeMethodOnSuperN(ScriptBytecodeAdapter.java:144)
 	at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy:142)
 	at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy)
-	at apitest.DeleteBookingTestApi.InvalidDelete(DeleteBookingTest.java:34)
+	at apitest.ReleaseDevice.InvalidDelete(DeleteBookingTest.java:34)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:78)
 	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
@@ -180,7 +180,7 @@ Expected status code &lt;405&gt; but was &lt;403&gt;.
 	at org.codehaus.groovy.runtime.ScriptBytecodeAdapter.invokeMethodOnSuperN(ScriptBytecodeAdapter.java:144)
 	at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy:142)
 	at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy)
-	at apitest.DeleteBookingTestApi.InvalidDelete(DeleteBookingTest.java:34)
+	at apitest.ReleaseDevice.InvalidDelete(DeleteBookingTest.java:34)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:78)
 	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
@@ -211,9 +211,9 @@ Expected status code &lt;405&gt; but was &lt;403&gt;.
 	at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:77)
 </pre></div></td>
 <td>1</td>
-<td>apitest.DeleteBookingTestApi@6da21078</td></tr>
+<td>apitest.ReleaseDevice@6da21078</td></tr>
 <tr>
-<td title='apitest.DeleteBookingTestApi.ValidDelete()'><b>ValidDelete</b><br>Test class: apitest.DeleteBookingTestApi</td>
+<td title='apitest.ReleaseDevice.ValidDelete()'><b>ValidDelete</b><br>Test class: apitest.ReleaseDevice</td>
 <td><div><pre>java.lang.AssertionError: 1 expectation failed.
 Expected status code &lt;201&gt; but was &lt;403&gt;.
 
@@ -261,7 +261,7 @@ Expected status code &lt;201&gt; but was &lt;403&gt;.
 	at org.codehaus.groovy.runtime.ScriptBytecodeAdapter.invokeMethodOnSuperN(ScriptBytecodeAdapter.java:144)
 	at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy:142)
 	at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy)
-	at apitest.DeleteBookingTestApi.ValidDelete(DeleteBookingTest.java:20)
+	at apitest.ReleaseDevice.ValidDelete(DeleteBookingTest.java:20)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:78)
 	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
@@ -338,7 +338,7 @@ Expected status code &lt;201&gt; but was &lt;403&gt;.
 	at org.codehaus.groovy.runtime.ScriptBytecodeAdapter.invokeMethodOnSuperN(ScriptBytecodeAdapter.java:144)
 	at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy:142)
 	at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy)
-	at apitest.DeleteBookingTestApi.ValidDelete(DeleteBookingTest.java:20)
+	at apitest.ReleaseDevice.ValidDelete(DeleteBookingTest.java:20)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:78)
 	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
@@ -369,7 +369,7 @@ Expected status code &lt;201&gt; but was &lt;403&gt;.
 	at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:77)
 </pre></div></td>
 <td>2</td>
-<td>apitest.DeleteBookingTestApi@6da21078</td></tr>
+<td>apitest.ReleaseDevice@6da21078</td></tr>
 <tr>
 <td title='apitest.UpdateBookingTestApi.ValidUpdate()'><b>ValidUpdate</b><br>Test class: apitest.UpdateBookingTestApi</td>
 <td><div><pre>java.lang.AssertionError: 1 expectation failed.
@@ -517,20 +517,20 @@ Expected status code &lt;200&gt; but was &lt;403&gt;.
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='apitest.GetBookingTestApi.Getbooking()'><b>Getbooking</b><br>Test class: apitest.GetBookingTestApi</td>
+<td title='apitest.GetNumberOfDevices.Getbooking()'><b>Getbooking</b><br>Test class: apitest.GetNumberOfDevices</td>
 <td></td>
 <td>1</td>
-<td>apitest.GetBookingTestApi@4229bb3f</td></tr>
+<td>apitest.GetNumberOfDevices@4229bb3f</td></tr>
 <tr>
 <td title='apitest.GenerateTokenTestApi.Token()'><b>Token</b><br>Test class: apitest.GenerateTokenTestApi</td>
 <td></td>
 <td>1</td>
 <td>apitest.GenerateTokenTestApi@2b91004a</td></tr>
 <tr>
-<td title='apitest.CreateBookingTestApi.ValidCreateBooking()'><b>ValidCreateBooking</b><br>Test class: apitest.CreateBookingTestApi</td>
+<td title='apitest.captureScreenShot.ValidCreateBooking()'><b>ValidCreateBooking</b><br>Test class: apitest.captureScreenShot</td>
 <td></td>
 <td>1</td>
-<td>apitest.CreateBookingTestApi@7fee8714</td></tr>
+<td>apitest.captureScreenShot@7fee8714</td></tr>
 </table><p>
 </body>
 </html>
