+++ test-output/Default suite/Default test.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
 <testsuite ignored="0" hostname="IN-PG02NMX7" failures="3" tests="6" name="Default apitest" time="10.364" errors="0" timestamp="2022-02-05T09:58:39 IST">
-  <testcase classname="apitest.DeleteBookingTestApi" name="ValidDelete" time="2.258">
+  <testcase classname="apitest.ReleaseDevice" name="ValidDelete" time="2.258">
     <failure type="java.lang.AssertionError" message="1 expectation failed.
 Expected status code &amp;lt;201&amp;gt; but was &amp;lt;403&amp;gt;.
 ">
@@ -52,7 +52,7 @@ at groovy.lang.MetaClassImpl.invokeMethod(MetaClassImpl.java:1268)
 at org.codehaus.groovy.runtime.ScriptBytecodeAdapter.invokeMethodOnSuperN(ScriptBytecodeAdapter.java:144)
 at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy:142)
 at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy)
-at apitest.DeleteBookingTestApi.ValidDelete(DeleteBookingTest.java:20)
+at apitest.ReleaseDevice.ValidDelete(DeleteBookingTest.java:20)
 at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:78)
 at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
@@ -84,7 +84,7 @@ at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:77)
 ]]>
     </failure>
   </testcase> <!-- ValidDelete -->
-  <testcase classname="apitest.DeleteBookingTestApi" name="InvalidDelete" time="1.417">
+  <testcase classname="apitest.ReleaseDevice" name="InvalidDelete" time="1.417">
     <failure type="java.lang.AssertionError" message="1 expectation failed.
 Expected status code &amp;lt;405&amp;gt; but was &amp;lt;403&amp;gt;.
 ">
@@ -125,7 +125,7 @@ at groovy.lang.MetaClassImpl.invokeMethod(MetaClassImpl.java:1268)
 at org.codehaus.groovy.runtime.ScriptBytecodeAdapter.invokeMethodOnSuperN(ScriptBytecodeAdapter.java:144)
 at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy:142)
 at io.restassured.internal.ValidatableResponseImpl.statusCode(ValidatableResponseImpl.groovy)
-at apitest.DeleteBookingTestApi.InvalidDelete(DeleteBookingTest.java:34)
+at apitest.ReleaseDevice.InvalidDelete(DeleteBookingTest.java:34)
 at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:78)
 at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
@@ -157,8 +157,8 @@ at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:77)
 ]]>
     </failure>
   </testcase> <!-- InvalidDelete -->
-  <testcase classname="apitest.CreateBookingTestApi" name="ValidCreateBooking" time="1.737"/>
-  <testcase classname="apitest.GetBookingTestApi" name="Getbooking" time="1.333"/>
+  <testcase classname="apitest.captureScreenShot" name="ValidCreateBooking" time="1.737"/>
+  <testcase classname="apitest.GetNumberOfDevices" name="Getbooking" time="1.333"/>
   <testcase classname="apitest.UpdateBookingTestApi" name="ValidUpdate" time="1.533">
     <failure type="java.lang.AssertionError" message="1 expectation failed.
 Expected status code &amp;lt;200&amp;gt; but was &amp;lt;403&amp;gt;.
