+++ test-output/Suite/Test.html
@@ -75,30 +75,30 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='apitest.GetBookingTestApi.Getbooking()'><b>Getbooking</b><br>Test class: apitest.GetBookingTestApi</td>
+<td title='apitest.GetNumberOfDevices.Getbooking()'><b>Getbooking</b><br>Test class: apitest.GetNumberOfDevices</td>
 <td></td>
 <td>1</td>
-<td>apitest.GetBookingTestApi@470734c3</td></tr>
+<td>apitest.GetNumberOfDevices@470734c3</td></tr>
 <tr>
-<td title='apitest.DeleteBookingTestApi.InvalidDelete()'><b>InvalidDelete</b><br>Test class: apitest.DeleteBookingTestApi</td>
+<td title='apitest.ReleaseDevice.InvalidDelete()'><b>InvalidDelete</b><br>Test class: apitest.ReleaseDevice</td>
 <td></td>
 <td>1</td>
-<td>apitest.DeleteBookingTestApi@2cf3d63b</td></tr>
+<td>apitest.ReleaseDevice@2cf3d63b</td></tr>
 <tr>
 <td title='apitest.GenerateTokenTestApi.Token()'><b>Token</b><br>Test class: apitest.GenerateTokenTestApi</td>
 <td></td>
 <td>2</td>
 <td>apitest.GenerateTokenTestApi@4b741d6d</td></tr>
 <tr>
-<td title='apitest.CreateBookingTestApi.ValidCreateBooking()'><b>ValidCreateBooking</b><br>Test class: apitest.CreateBookingTestApi</td>
+<td title='apitest.captureScreenShot.ValidCreateBooking()'><b>ValidCreateBooking</b><br>Test class: apitest.captureScreenShot</td>
 <td></td>
 <td>1</td>
-<td>apitest.CreateBookingTestApi@2eae8e6e</td></tr>
+<td>apitest.captureScreenShot@2eae8e6e</td></tr>
 <tr>
-<td title='apitest.DeleteBookingTestApi.ValidDelete()'><b>ValidDelete</b><br>Test class: apitest.DeleteBookingTestApi</td>
+<td title='apitest.ReleaseDevice.ValidDelete()'><b>ValidDelete</b><br>Test class: apitest.ReleaseDevice</td>
 <td></td>
 <td>1</td>
-<td>apitest.DeleteBookingTestApi@2cf3d63b</td></tr>
+<td>apitest.ReleaseDevice@2cf3d63b</td></tr>
 <tr>
 <td title='apitest.UpdateBookingTestApi.ValidUpdate()'><b>ValidUpdate</b><br>Test class: apitest.UpdateBookingTestApi</td>
 <td></td>
