+++ test-output/Suite/Test.xml
@@ -2,39 +2,39 @@
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
 <testsuite ignored="10" hostname="IN-PG02NMX7" failures="0" tests="6" name="Test" time="10.784" errors="0" timestamp="2022-02-07T20:28:10 IST">
   <testcase classname="apitest.GenerateTokenTestApi" name="Token" time="2.914"/>
-  <testcase classname="apitest.CreateBookingTestApi" name="ValidCreateBooking" time="1.781"/>
+  <testcase classname="apitest.captureScreenShot" name="ValidCreateBooking" time="1.781"/>
   <testcase classname="apitest.UpdateBookingTestApi" name="ValidUpdate" time="1.514"/>
-  <testcase classname="apitest.GetBookingTestApi" name="Getbooking" time="1.381"/>
-  <testcase classname="apitest.DeleteBookingTestApi" name="ValidDelete" time="1.372"/>
-  <testcase classname="apitest.DeleteBookingTestApi" name="InvalidDelete" time="1.416"/>
-  <testcase classname="baseclass.ApiBaseTest" name="@BeforeSuite setExtent" time="0.0">
+  <testcase classname="apitest.GetNumberOfDevices" name="Getbooking" time="1.381"/>
+  <testcase classname="apitest.ReleaseDevice" name="ValidDelete" time="1.372"/>
+  <testcase classname="apitest.ReleaseDevice" name="InvalidDelete" time="1.416"/>
+  <testcase classname="baseclass.ApiBase" name="@BeforeSuite setExtent" time="0.0">
     <ignored/>
   </testcase> <!-- @BeforeSuite setExtent -->
-  <testcase classname="baseclass.ApiBaseTest" name="@BeforeSuite setExtent" time="0.0">
+  <testcase classname="baseclass.ApiBase" name="@BeforeSuite setExtent" time="0.0">
     <ignored/>
   </testcase> <!-- @BeforeSuite setExtent -->
-  <testcase classname="baseclass.ApiBaseTest" name="@BeforeSuite setExtent" time="0.0">
+  <testcase classname="baseclass.ApiBase" name="@BeforeSuite setExtent" time="0.0">
     <ignored/>
   </testcase> <!-- @BeforeSuite setExtent -->
-  <testcase classname="baseclass.ApiBaseTest" name="@BeforeSuite setExtent" time="0.0">
+  <testcase classname="baseclass.ApiBase" name="@BeforeSuite setExtent" time="0.0">
     <ignored/>
   </testcase> <!-- @BeforeSuite setExtent -->
-  <testcase classname="baseclass.ApiBaseTest" name="@BeforeSuite setExtent" time="0.0">
+  <testcase classname="baseclass.ApiBase" name="@BeforeSuite setExtent" time="0.0">
     <ignored/>
   </testcase> <!-- @BeforeSuite setExtent -->
-  <testcase classname="baseclass.ApiBaseTest" name="@AfterSuite endReport" time="0.0">
+  <testcase classname="baseclass.ApiBase" name="@AfterSuite endReport" time="0.0">
     <ignored/>
   </testcase> <!-- @AfterSuite endReport -->
-  <testcase classname="baseclass.ApiBaseTest" name="@AfterSuite endReport" time="0.0">
+  <testcase classname="baseclass.ApiBase" name="@AfterSuite endReport" time="0.0">
     <ignored/>
   </testcase> <!-- @AfterSuite endReport -->
-  <testcase classname="baseclass.ApiBaseTest" name="@AfterSuite endReport" time="0.0">
+  <testcase classname="baseclass.ApiBase" name="@AfterSuite endReport" time="0.0">
     <ignored/>
   </testcase> <!-- @AfterSuite endReport -->
-  <testcase classname="baseclass.ApiBaseTest" name="@AfterSuite endReport" time="0.0">
+  <testcase classname="baseclass.ApiBase" name="@AfterSuite endReport" time="0.0">
     <ignored/>
   </testcase> <!-- @AfterSuite endReport -->
-  <testcase classname="baseclass.ApiBaseTest" name="@AfterSuite endReport" time="0.0">
+  <testcase classname="baseclass.ApiBase" name="@AfterSuite endReport" time="0.0">
     <ignored/>
   </testcase> <!-- @AfterSuite endReport -->
 </testsuite> <!-- Test -->
