+++ test-output/Suite/testng-failed.xml
@@ -3,7 +3,7 @@
 <suite name="Failed suite [Suite]" guice-stage="DEVELOPMENT">
   <test thread-count="5" name="Test(failed)">
     <classes>
-      <class name="apitest.CreateBookingTestApi">
+      <class name="apitest.captureScreenShot">
         <methods>
           <include name="setExtent"/>
           <include name="attachScreenhsot"/>
@@ -12,8 +12,8 @@
           <include name="tearDown"/>
           <include name="ValidCreateBooking"/>
         </methods>
-      </class> <!-- apitest.CreateBookingTestApi -->
-      <class name="apitest.DeleteBookingTestApi">
+      </class> <!-- apitest.captureScreenShot -->
+      <class name="apitest.ReleaseDevice">
         <methods>
           <include name="setExtent"/>
           <include name="attachScreenhsot"/>
@@ -22,7 +22,7 @@
           <include name="setUp"/>
           <include name="tearDown"/>
         </methods>
-      </class> <!-- apitest.DeleteBookingTestApi -->
+      </class> <!-- apitest.ReleaseDevice -->
     </classes>
   </test> <!-- Test(failed) -->
 </suite> <!-- Failed suite [Suite] -->
