+++ test-output/emailable-report.html
@@ -11,13 +11,13 @@
 <tr><th colspan="8">Suite</th></tr>
 <tr><td><a href="#t0">Test</a></td><td class="num">6</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">10,784</td><td></td><td></td></tr>
 </table>
-<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="1">apitest.CreateBookingTestApi</td><td><a href="#m0">ValidCreateBooking</a></td><td rowspan="1">1644245882785</td><td rowspan="1">1781</td></tr><tr class="passedodd"><td rowspan="2">apitest.DeleteBookingTestApi</td><td><a href="#m1">InvalidDelete</a></td><td rowspan="1">1644245888845</td><td rowspan="1">1416</td></tr><tr class="passedodd"><td><a href="#m2">ValidDelete</a></td><td rowspan="1">1644245887471</td><td rowspan="1">1372</td></tr><tr class="passedeven"><td rowspan="1">apitest.GenerateTokenTestApi</td><td><a href="#m3">Token</a></td><td rowspan="1">1644245879865</td><td rowspan="1">2914</td></tr><tr class="passedodd"><td rowspan="1">apitest.GetBookingTestApi</td><td><a href="#m4">Getbooking</a></td><td rowspan="1">1644245886087</td><td rowspan="1">1381</td></tr><tr class="passedeven"><td rowspan="1">apitest.UpdateBookingTestApi</td><td><a href="#m5">ValidUpdate</a></td><td rowspan="1">1644245884571</td><td rowspan="1">1514</td></tr></tbody>
+<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="1">apitest.captureScreenShot</td><td><a href="#m0">ValidCreateBooking</a></td><td rowspan="1">1644245882785</td><td rowspan="1">1781</td></tr><tr class="passedodd"><td rowspan="2">apitest.ReleaseDevice</td><td><a href="#m1">InvalidDelete</a></td><td rowspan="1">1644245888845</td><td rowspan="1">1416</td></tr><tr class="passedodd"><td><a href="#m2">ValidDelete</a></td><td rowspan="1">1644245887471</td><td rowspan="1">1372</td></tr><tr class="passedeven"><td rowspan="1">apitest.GenerateTokenTestApi</td><td><a href="#m3">Token</a></td><td rowspan="1">1644245879865</td><td rowspan="1">2914</td></tr><tr class="passedodd"><td rowspan="1">apitest.GetNumberOfDevices</td><td><a href="#m4">Getbooking</a></td><td rowspan="1">1644245886087</td><td rowspan="1">1381</td></tr><tr class="passedeven"><td rowspan="1">apitest.UpdateBookingTestApi</td><td><a href="#m5">ValidUpdate</a></td><td rowspan="1">1644245884571</td><td rowspan="1">1514</td></tr></tbody>
 </table>
-<h2>Test</h2><h3 id="m0">apitest.CreateBookingTestApi#ValidCreateBooking</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m1">apitest.DeleteBookingTestApi#InvalidDelete</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m2">apitest.DeleteBookingTestApi#ValidDelete</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h2>Test</h2><h3 id="m0">apitest.captureScreenShot#ValidCreateBooking</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m1">apitest.ReleaseDevice#InvalidDelete</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m2">apitest.ReleaseDevice#ValidDelete</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 <h3 id="m3">apitest.GenerateTokenTestApi#Token</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m4">apitest.GetBookingTestApi#Getbooking</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m4">apitest.GetNumberOfDevices#Getbooking</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 <h3 id="m5">apitest.UpdateBookingTestApi#ValidUpdate</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 </body>
 </html>
