+++ test-output/junitreports/TEST-test.BaseTest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="IN-PG02NMX7" failures="0" tests="1" name="baseclass.ApiBaseTest" time="2.723" errors="0" timestamp="2022-02-03T17:32:33 IST" skipped="0">
-  <testcase classname="baseclass.ApiBaseTest" name="pingTest" time="2.723"/>
+<testsuite hostname="IN-PG02NMX7" failures="0" tests="1" name="baseclass.ApiBase" time="2.723" errors="0" timestamp="2022-02-03T17:32:33 IST" skipped="0">
+  <testcase classname="baseclass.ApiBase" name="pingTest" time="2.723"/>
   <system-out/>
-</testsuite> <!-- baseclass.ApiBaseTest -->
+</testsuite> <!-- baseclass.ApiBase -->
