+++ test-output/junitreports/TEST-test.CreateBookingTest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="IN-PG02NMX7" failures="0" tests="1" name="apitest.CreateBookingTestApi" time="1.782" errors="0" timestamp="2022-02-07T20:28:10 IST" skipped="0">
-  <testcase classname="apitest.CreateBookingTestApi" name="ValidCreateBooking" time="1.782"/>
+<testsuite hostname="IN-PG02NMX7" failures="0" tests="1" name="apitest.captureScreenShot" time="1.782" errors="0" timestamp="2022-02-07T20:28:10 IST" skipped="0">
+  <testcase classname="apitest.captureScreenShot" name="ValidCreateBooking" time="1.782"/>
   <system-out/>
-</testsuite> <!-- apitest.CreateBookingTestApi -->
+</testsuite> <!-- apitest.captureScreenShot -->
