+++ test-output/junitreports/TEST-test.DeleteBookingTest.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="IN-PG02NMX7" failures="0" tests="2" name="apitest.DeleteBookingTestApi" time="2.790" errors="0" timestamp="2022-02-07T20:28:10 IST" skipped="0">
-  <testcase classname="apitest.DeleteBookingTestApi" name="ValidDelete" time="1.373"/>
+<testsuite hostname="IN-PG02NMX7" failures="0" tests="2" name="apitest.ReleaseDevice" time="2.790" errors="0" timestamp="2022-02-07T20:28:10 IST" skipped="0">
+  <testcase classname="apitest.ReleaseDevice" name="ValidDelete" time="1.373"/>
   <system-out/>
-  <testcase classname="apitest.DeleteBookingTestApi" name="InvalidDelete" time="1.417"/>
+  <testcase classname="apitest.ReleaseDevice" name="InvalidDelete" time="1.417"/>
   <system-out/>
-</testsuite> <!-- apitest.DeleteBookingTestApi -->
+</testsuite> <!-- apitest.ReleaseDevice -->
