+++ test-output/junitreports/TEST-test.GetBookingTest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="IN-PG02NMX7" failures="0" tests="1" name="apitest.GetBookingTestApi" time="1.381" errors="0" timestamp="2022-02-07T20:28:10 IST" skipped="0">
-  <testcase classname="apitest.GetBookingTestApi" name="Getbooking" time="1.381"/>
+<testsuite hostname="IN-PG02NMX7" failures="0" tests="1" name="apitest.GetNumberOfDevices" time="1.381" errors="0" timestamp="2022-02-07T20:28:10 IST" skipped="0">
+  <testcase classname="apitest.GetNumberOfDevices" name="Getbooking" time="1.381"/>
   <system-out/>
-</testsuite> <!-- apitest.GetBookingTestApi -->
+</testsuite> <!-- apitest.GetNumberOfDevices -->
