+++ test-output/old/Default suite/groups.html
@@ -1,6 +1,6 @@
 <h2>Groups used for this apitest run</h2><table border="1">
-<tr> <td align="center"><b>Group name</b></td><td align="center"><b>Methods</b></td></tr><tr><td>delete</td><td>DeleteBookingTest.InvalidDelete()[pri:6, instance:apitest.DeleteBookingTestApi@6da21078]<br/>DeleteBookingTest.ValidDelete()[pri:5, instance:apitest.DeleteBookingTestApi@6da21078]<br/></td></tr>
-<tr><td>get</td><td>GetBookingTest.Getbooking()[pri:4, instance:apitest.GetBookingTestApi@4229bb3f]<br/></td></tr>
-<tr><td>post</td><td>GenerateTokenTest.Token()[pri:1, instance:apitest.GenerateTokenTestApi@2b91004a]<br/>CreateBookingTest.ValidCreateBooking()[pri:2, instance:apitest.CreateBookingTestApi@7fee8714]<br/></td></tr>
+<tr> <td align="center"><b>Group name</b></td><td align="center"><b>Methods</b></td></tr><tr><td>delete</td><td>DeleteBookingTest.InvalidDelete()[pri:6, instance:apitest.ReleaseDevice@6da21078]<br/>DeleteBookingTest.ValidDelete()[pri:5, instance:apitest.ReleaseDevice@6da21078]<br/></td></tr>
+<tr><td>get</td><td>GetBookingTest.Getbooking()[pri:4, instance:apitest.GetNumberOfDevices@4229bb3f]<br/></td></tr>
+<tr><td>post</td><td>GenerateTokenTest.Token()[pri:1, instance:apitest.GenerateTokenTestApi@2b91004a]<br/>CreateBookingTest.ValidCreateBooking()[pri:2, instance:apitest.captureScreenShot@7fee8714]<br/></td></tr>
 <tr><td>put</td><td>UpdateBookingTest.ValidUpdate()[pri:3, instance:apitest.UpdateBookingTestApi@56cdfb3b]<br/></td></tr>
 </table>
