+++ test-output/old/Default suite/methods.html
@@ -1,29 +1,29 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the apitest class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="677d7f">  <td>22/02/05 09:58:29</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseTest.setUp()[pri:0, instance:apitest.DeleteBookingTestApi@6da21078]">&gt;&gt;setUp</td>
+<tr bgcolor="677d7f">  <td>22/02/05 09:58:29</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseTest.setUp()[pri:0, instance:apitest.ReleaseDevice@6da21078]">&gt;&gt;setUp</td>
 <td>&nbsp;</td>  <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="7b6572">  <td>22/02/05 09:58:29</td>   <td>423</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="DeleteBookingTest.ValidDelete()[pri:5, instance:apitest.DeleteBookingTestApi@6da21078]">ValidDelete</td>
+<tr bgcolor="7b6572">  <td>22/02/05 09:58:29</td>   <td>423</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="DeleteBookingTest.ValidDelete()[pri:5, instance:apitest.ReleaseDevice@6da21078]">ValidDelete</td>
   <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="677d7f">  <td>22/02/05 09:58:31</td>   <td>2684</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseTest.tearDown()[pri:0, instance:apitest.DeleteBookingTestApi@6da21078]">&lt;&lt;tearDown</td>
+<tr bgcolor="677d7f">  <td>22/02/05 09:58:31</td>   <td>2684</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseTest.tearDown()[pri:0, instance:apitest.ReleaseDevice@6da21078]">&lt;&lt;tearDown</td>
 <td>&nbsp;</td>  <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="677d7f">  <td>22/02/05 09:58:31</td>   <td>2686</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseTest.setUp()[pri:0, instance:apitest.DeleteBookingTestApi@6da21078]">&gt;&gt;setUp</td>
+<tr bgcolor="677d7f">  <td>22/02/05 09:58:31</td>   <td>2686</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseTest.setUp()[pri:0, instance:apitest.ReleaseDevice@6da21078]">&gt;&gt;setUp</td>
 <td>&nbsp;</td>  <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="7b6572">  <td>22/02/05 09:58:31</td>   <td>2687</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="DeleteBookingTest.InvalidDelete()[pri:6, instance:apitest.DeleteBookingTestApi@6da21078]">InvalidDelete</td>
+<tr bgcolor="7b6572">  <td>22/02/05 09:58:31</td>   <td>2687</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="DeleteBookingTest.InvalidDelete()[pri:6, instance:apitest.ReleaseDevice@6da21078]">InvalidDelete</td>
   <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="677d7f">  <td>22/02/05 09:58:33</td>   <td>4106</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseTest.tearDown()[pri:0, instance:apitest.DeleteBookingTestApi@6da21078]">&lt;&lt;tearDown</td>
+<tr bgcolor="677d7f">  <td>22/02/05 09:58:33</td>   <td>4106</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseTest.tearDown()[pri:0, instance:apitest.ReleaseDevice@6da21078]">&lt;&lt;tearDown</td>
 <td>&nbsp;</td>  <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="677d7f">  <td>22/02/05 09:58:33</td>   <td>4106</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseTest.setUp()[pri:0, instance:apitest.CreateBookingTestApi@7fee8714]">&gt;&gt;setUp</td>
+<tr bgcolor="677d7f">  <td>22/02/05 09:58:33</td>   <td>4106</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseTest.setUp()[pri:0, instance:apitest.captureScreenShot@7fee8714]">&gt;&gt;setUp</td>
 <td>&nbsp;</td>  <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="977460">  <td>22/02/05 09:58:33</td>   <td>4107</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="CreateBookingTest.ValidCreateBooking()[pri:2, instance:apitest.CreateBookingTestApi@7fee8714]">ValidCreateBooking</td>
+<tr bgcolor="977460">  <td>22/02/05 09:58:33</td>   <td>4107</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="CreateBookingTest.ValidCreateBooking()[pri:2, instance:apitest.captureScreenShot@7fee8714]">ValidCreateBooking</td>
   <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="677d7f">  <td>22/02/05 09:58:35</td>   <td>5845</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseTest.tearDown()[pri:0, instance:apitest.CreateBookingTestApi@7fee8714]">&lt;&lt;tearDown</td>
+<tr bgcolor="677d7f">  <td>22/02/05 09:58:35</td>   <td>5845</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseTest.tearDown()[pri:0, instance:apitest.captureScreenShot@7fee8714]">&lt;&lt;tearDown</td>
 <td>&nbsp;</td>  <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="677d7f">  <td>22/02/05 09:58:35</td>   <td>5846</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseTest.setUp()[pri:0, instance:apitest.GetBookingTestApi@4229bb3f]">&gt;&gt;setUp</td>
+<tr bgcolor="677d7f">  <td>22/02/05 09:58:35</td>   <td>5846</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseTest.setUp()[pri:0, instance:apitest.GetNumberOfDevices@4229bb3f]">&gt;&gt;setUp</td>
 <td>&nbsp;</td>  <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="b7cd72">  <td>22/02/05 09:58:35</td>   <td>5846</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="GetBookingTest.Getbooking()[pri:4, instance:apitest.GetBookingTestApi@4229bb3f]">Getbooking</td>
+<tr bgcolor="b7cd72">  <td>22/02/05 09:58:35</td>   <td>5846</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="GetBookingTest.Getbooking()[pri:4, instance:apitest.GetNumberOfDevices@4229bb3f]">Getbooking</td>
   <td>main@2067180044</td>   <td></td> </tr>
-<tr bgcolor="677d7f">  <td>22/02/05 09:58:36</td>   <td>7180</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseTest.tearDown()[pri:0, instance:apitest.GetBookingTestApi@4229bb3f]">&lt;&lt;tearDown</td>
+<tr bgcolor="677d7f">  <td>22/02/05 09:58:36</td>   <td>7180</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseTest.tearDown()[pri:0, instance:apitest.GetNumberOfDevices@4229bb3f]">&lt;&lt;tearDown</td>
 <td>&nbsp;</td>  <td>main@2067180044</td>   <td></td> </tr>
 <tr bgcolor="677d7f">  <td>22/02/05 09:58:36</td>   <td>7181</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseTest.setUp()[pri:0, instance:apitest.UpdateBookingTestApi@56cdfb3b]">&gt;&gt;setUp</td>
 <td>&nbsp;</td>  <td>main@2067180044</td>   <td></td> </tr>
