+++ test-output/old/Suite/groups.html
@@ -1,6 +1,6 @@
 <h2>Groups used for this apitest run</h2><table border="1">
-<tr> <td align="center"><b>Group name</b></td><td align="center"><b>Methods</b></td></tr><tr><td>delete</td><td>DeleteBookingTest.InvalidDelete()[pri:6, instance:apitest.DeleteBookingTestApi@2cf3d63b]<br/>DeleteBookingTest.ValidDelete()[pri:5, instance:apitest.DeleteBookingTestApi@2cf3d63b]<br/></td></tr>
-<tr><td>get</td><td>GetBookingTest.Getbooking()[pri:4, instance:apitest.GetBookingTestApi@470734c3]<br/></td></tr>
-<tr><td>post</td><td>GenerateTokenTest.Token()[pri:1, instance:apitest.GenerateTokenTestApi@4b741d6d]<br/>CreateBookingTest.ValidCreateBooking()[pri:2, instance:apitest.CreateBookingTestApi@2eae8e6e]<br/></td></tr>
+<tr> <td align="center"><b>Group name</b></td><td align="center"><b>Methods</b></td></tr><tr><td>delete</td><td>DeleteBookingTest.InvalidDelete()[pri:6, instance:apitest.ReleaseDevice@2cf3d63b]<br/>DeleteBookingTest.ValidDelete()[pri:5, instance:apitest.ReleaseDevice@2cf3d63b]<br/></td></tr>
+<tr><td>get</td><td>GetBookingTest.Getbooking()[pri:4, instance:apitest.GetNumberOfDevices@470734c3]<br/></td></tr>
+<tr><td>post</td><td>GenerateTokenTest.Token()[pri:1, instance:apitest.GenerateTokenTestApi@4b741d6d]<br/>CreateBookingTest.ValidCreateBooking()[pri:2, instance:apitest.captureScreenShot@2eae8e6e]<br/></td></tr>
 <tr><td>put</td><td>UpdateBookingTest.ValidUpdate()[pri:3, instance:apitest.UpdateBookingTestApi@8f2ef19]<br/></td></tr>
 </table>
