+++ test-output/old/Suite/methods-not-run.html
@@ -1,12 +1,12 @@
 <h2>Methods that were not run</h2><table>
-<tr><td>baseclass.ApiBaseTest.setExtent</td></tr>
-<tr><td>baseclass.ApiBaseTest.setExtent</td></tr>
-<tr><td>baseclass.ApiBaseTest.setExtent</td></tr>
-<tr><td>baseclass.ApiBaseTest.setExtent</td></tr>
-<tr><td>baseclass.ApiBaseTest.setExtent</td></tr>
-<tr><td>baseclass.ApiBaseTest.endReport</td></tr>
-<tr><td>baseclass.ApiBaseTest.endReport</td></tr>
-<tr><td>baseclass.ApiBaseTest.endReport</td></tr>
-<tr><td>baseclass.ApiBaseTest.endReport</td></tr>
-<tr><td>baseclass.ApiBaseTest.endReport</td></tr>
+<tr><td>baseclass.ApiBase.setExtent</td></tr>
+<tr><td>baseclass.ApiBase.setExtent</td></tr>
+<tr><td>baseclass.ApiBase.setExtent</td></tr>
+<tr><td>baseclass.ApiBase.setExtent</td></tr>
+<tr><td>baseclass.ApiBase.setExtent</td></tr>
+<tr><td>baseclass.ApiBase.endReport</td></tr>
+<tr><td>baseclass.ApiBase.endReport</td></tr>
+<tr><td>baseclass.ApiBase.endReport</td></tr>
+<tr><td>baseclass.ApiBase.endReport</td></tr>
+<tr><td>baseclass.ApiBase.endReport</td></tr>
 </table>
