+++ testng.xml
@@ -3,12 +3,9 @@
 <suite name="Suite" >
   <test thread-count="5" name="Test" >
     <classes>
-      <class name="baseclass.ApiBaseTest"/>
-      <class name="apitest.GenerateTokenTestApi"/>
-      <class name="apitest.CreateBookingTestApi"/>
-      <class name="apitest.UpdateBookingTestApi"/>
-      <class name="apitest.GetBookingTestApi"/>
-      <class name="apitest.DeleteBookingTestApi"/>
+      <class name="apitest.GetNumberOfDevices"/>
+      <class name="apitest.CaptureScreenShot"/>
+      <class name="apitest.ReleaseDevice"/>
     </classes>
   </test> <!-- Test -->
 </suite> <!-- Suite -->
